+++ noobOsu.Game/Beatmaps/Timing/IBeatmapTiming.cs
@@ -4,14 +4,14 @@ namespace noobOsu.Game.Beatmaps.Timing
 {
     public interface IBeatmapTiming
     {
-        float BPM_average { get; }
-        float BPM_min { get; }
-        float BPM_max { get; }
-        float FirstBPM { get; }
+        double BPM_average { get; }
+        double BPM_min { get; }
+        double BPM_max { get; }
+        double FirstBPM { get; }
         
         ITimingPoint GetTimingPoint(int timestamp);
         ITimingPoint GetTimingPoint(int timestamp, bool uninherited);
-        float BPM_At(int timestamp);
+        double BPM_At(int timestamp);
         IReadOnlyList<ITimingPoint> GetTimingPoints();
     }
 }
