+++ noobOsu.Game/HitObjects/Drawables/HitObjectSprite.cs
@@ -4,6 +4,6 @@ namespace noobOsu.Game.HitObjects.Drawables
 {
     public partial class HitObjectSprite : Sprite
     {
-        public float ScaleFactor { get; set; } = 1f;
+        public double ScaleFactor { get; set; } = 1d;
     }
 }
